fix(ui): break vault settings circular dependency and hard-fail cycles in desktop build
