+++ core/ui/vault/settings/backup/index.tsx
@@ -6,7 +6,10 @@ import { useTranslation } from 'react-i18next'
 
 import { useCoreNavigate } from '../../../navigation/hooks/useCoreNavigate'
 import { useCurrentVault } from '../../state/currentVault'
-import { DescriptionText, ListItemIconWrapper } from '..'
+import {
+  DescriptionText,
+  ListItemIconWrapper,
+} from '../vaultSettingsListStyles'
 import { BackupModal } from './BackupModal'
 import { backupOptionView } from './routes'
